+++ src/client/java/online/connlost/ConfigSync.java
@@ -2,19 +2,18 @@ package online.connlost;
 
 import online.connlost.server.config.ConfigManager;
 import online.connlost.util.ItemsHelper;
-import online.connlost.AllStackable;
 
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
 public class ConfigSync {
-    private static ItemsHelper itemsHelper = ItemsHelper.getItemsHelper();
+    private static final ItemsHelper itemsHelper = ItemsHelper.getItemsHelper();
 
     /**
      * Items' stacksize is directly modified, no config is held for that in the "client" side.
      * In contrast, we do store rules in client side as some mixins in client side classes need it.
-     * @param configList
+     * @param configList the config list
      */
     public static void syncConfig(ArrayList<LinkedHashMap<String, Integer>> configList){
         AllStackable.LOGGER.info("[All Stackable] [Client] Sync config from server side!");
@@ -22,13 +21,12 @@ public class ConfigSync {
         ConfigManager.getConfigManager().setRulesMap(configList.get(1));
         AllStackable.LOGGER.info("[All Stackable] [Client] Sync rules:");
         for (Map.Entry<String, Integer> rule: configList.get(1).entrySet()){
-            String tag;
-            switch (rule.getValue()){
-                case 0: tag = "false";break;
-                case 1: tag = "true";break;
-                default: tag = rule.getValue().toString();
-            }
-            AllStackable.LOGGER.info("\t["+rule.getKey()+"] = "+tag);
+            String tag = switch (rule.getValue()) {
+                case 0 -> "false";
+                case 1 -> "true";
+                default -> rule.getValue().toString();
+            };
+            AllStackable.LOGGER.info("\t[{}] = {}", rule.getKey(), tag);
         }
         AllStackable.LOGGER.info("[All Stackable] [Client] Sync finished.");
     }
